JIT: NestedJIT spaceship/exception ABI toward EnumSpaceship (#21124) (#21123)
* JIT: NestedJIT spaceship/exception ABI toward EnumSpaceship (#21109)

Pass numOnLeft as i64, type HashTable params as __hashtable__*, slim
CompareJitHelper NestedJIT surface (CompareUnlikeHelper), restore
JitStatPathKernel insert block, fix enum_from string phi dominance,
and align ExceptionHandlerJitHelper with empty-string NestedJIT ABI.

EnumSpaceshipJit still fails module verify ([8 x i8] icmp, value
bitcasts, preg_quote/?string, depth i32/i64) — handoff next.


* JIT: NestedJIT ABI + MCJIT mem bodies for EnumSpaceship (#21109)

Finish phantom getValue follow-ons: scalar spaceship NestedJIT, EH/PregQuote
ABI, GC trunc/registry, and module-local memset/memcpy/memcmp/strlen/strcmp
plus write→syscall so MCJIT __init__ no longer call-through-null. EnumSpaceship
compile/verify is green; execute still blocked on libc snprintf relocation.


* JIT: MCJIT embed echo via php_write + unit enum <=> (#21124)

Route EMBED echo through a bound php_write function pointer (ob-aware) and IR
itoa for int echo so LLVM 9 MCJIT no longer calls null libc snprintf/write.
Unit-enum spaceship returns 1 for distinct singletons; enum name compare uses
module-local memcmp instead of strncasecmp.


* Docs: resync bootstrap inventory after McjitEmbedHostEcho (#21124)


---------

+++ ext/standard/phpc_object_is_constructed_native.php
@@ -31,6 +31,7 @@ final class phpc_object_is_constructed_native extends Internal
         }
         $isConstructed = GcCollectCyclesNativeOpsJit::objectIsConstructed($context, $args[0]);
 
-        return $context->builder->zext($isConstructed, $context->getTypeFromString('int32'));
+        // NestedJIT boxes PHP ints via __value__writeLong (i64); i32 zext breaks module verify (#21109).
+        return $context->builder->zext($isConstructed, $context->getTypeFromString('int64'));
     }
 }
